user_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

# ORM & DB
SQLAlchemy==2.0.31
asyncpg==0.29.0   # PostgreSQL için async sürücü

# Pydantic (validation & settings)
pydantic==2.7.4
pydantic-settings==2.3.2

# Auth & Security
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
